fix fonts&data
