+++ src/components/common/profile/profile.data.js
@@ -1,0 +1,67 @@
+import { PROFILE_TEMPLATE_TYPES } from "constants/profile.constants";
+
+const profilesData = [
+  {
+    fullName: "Василевська Василина ",
+    id: "1",
+    description:`Тисячі любителів футболу та спорту мають причини бути вдячними моїй сестрі Василині, яка померла у віці 55 років від множинних ракових захворювань через дев'ять років після першого діагнозу рак молочної залози.
+
+      Як керівник судових справ та конституційних справ у Supporter Direct, організації, яка сприяє залученню вболівальників до управління їх клубами та заохочує їх створювати акціонерні фонди, вона допомогла створити майже 200 таких трестів по всій країні.
+
+      У 2017 році вона заснувала Жінки в грі, щоб заохотити більше жінок ходити на прямі матчі. Вона хотіла допомогти жінкам почуватись комфортно, відвідуючи матчі. Ініціатива розпочалася на нелігових іграх і швидко закріпилася в Прем'єр-лізі. У минулому сезоні "Манчестер Сіті" і "Хаддерсфілд Таун".......`,
+    templateType: PROFILE_TEMPLATE_TYPES.ARTICLE,
+    imgSrc: "",
+  },
+  {
+    fullName: "Альберт 'Гордон' Мюррей ",
+    id: "2",
+    description:
+      `Привіт там і ласкаво просимо на сайт родинних спогадів Мюррея.
+
+      Ця гілка родини Мюррей походить із західного кінця Ньюкасла на Тайн.
+      
+      Перше зображення, яке ви бачите, - Альберт "Гордон" Мюррей з його дочкою Лоїс Джоанн та його сином Гремем Нілом - це було зроблено десь у 1971 році після народження Грема.
+      
+      Гордон, як він був відомий, народився 9 вересня 1928 року в Ньюкаслі і разом з Морін Елізабет мав 3 дітей - Лоїс Джоан (1969-), Грем Ніл (1971-) і Маргарет Кей (1973-).
+      `,
+    templateType: PROFILE_TEMPLATE_TYPES.BOOK,
+    imgSrc: "",
+  },
+  {
+    fullName: "Алан Марк Хоббс ",
+    id: "3",
+    description:
+      `Алан народився 14 червня 1976 року і, на жаль, помер 27 лютого 2014 року у віці лише 37 років.
+
+      Завжди працьовитий працівник, Алан будував свій власний бізнес, починаючи від перевезення автомобілів до новаторських дрібних вм’ятин та дрібних компаній з ремонту кузовів, випереджаючи конкурентів на роки, випробувавши найновіше обладнання з усього світу.
+      
+      Алан надзвичайно любив свою музику, їздив на багато концертів, володів сотнями компакт-дисків і мав лише найкращі музичні системи, будь то вдома чи в одному зі своїх транспортних засобів, Pink Floyd був його улюбленим і навіть мав татуювання, щоб довести це!`,
+    templateType: PROFILE_TEMPLATE_TYPES.SIMPLE,
+    imgSrc: "",
+  },
+  {
+    fullName: "Міллісент Емілі Патрік (Milly) ",
+    id: "4",
+    description:
+      `Міллісент Патрік народилася Мілісент Емілі Пірс 19 грудня 1923 року. Вона була єдиною дочкою одного з трьох братів, у кожного з яких була одинока дочка і жодного іншого потомства. За цих обставин було цілком природно, що три дівчинки - Беатріче, Міллісент та Олена стали більше схожими на сестер, ніж на двоюрідних сестер. Для дітей Беатріче, Роджера та Джудіт, а також їхніх дітей та сімей, Міллісент завжди була тіткою Міллі.
+
+      Вперше сім'я знала Тетю Міллі, коли вона жила з матір'ю та батьком на Кейпстоун-роуд, Борнмут, у напівбудинку по сусідству з тіткою Ліл, сестрою її матері. Ці дві сім'ї були настільки близькі, що Роджер і Джудіт були в дитинстві захоплені тим, що цегла була розпушена в стіні партії і її можна було висунути, щоб вони могли спілкуватися між собою.`,
+    templateType: PROFILE_TEMPLATE_TYPES.BOOK,
+    imgSrc: "",
+  },
+  {
+    fullName: "Льюїс Дантон ",
+    id: "5",
+    description:
+      `Льюїс Дантон Серфер, футболіст, ді-джей і друг усім (особливо самотнім або сором'язливим) помер 10 лютого 2005 року.
+
+      Він втратив своє життя через отруєння чадним газом, забравши вугілля вечірнього багаття до свого намету, щоб зігрітися. Він був на пляжі в Марокко, де він займався серфінгом.
+      
+      Серфінг був життям Льюїса і охопив його по всьому світу. Футбол, крикет, музика та дівчата також вирізнялися сильно. Коли він не був у воді, його часто можна було десь знайти ді-джеєм. Під час поїздки додому з Австралії він мав зупинку в Південній Кореї, яка стала шестимісячним заклинанням для навчання маленьких корейців англійській мові.`,
+    templateType: PROFILE_TEMPLATE_TYPES.ARTICLE,
+    imgSrc: "",
+  },
+];
+
+
+export default profilesData;
