+++ src/components/profiles/profiles-card/style.scss
@@ -37,13 +37,13 @@
   }
 
   &__full-name {
-    font-family: 'Playfair Display';
+    font-family: $theme-f-playfair;
     font-size: 24px;
     line-height: 36px;
   }
 
   &__description {
-    font-family: 'Open Sans';
+    font-family: $theme-f-opensans;
     line-height: 24px;
     white-space: pre-line;
   }
