+++ src/router/routes.js
@@ -1,6 +1,7 @@
 import routesConstants from "constants/routes.constants";
 import HomePage from "pages/home";
 import { ProfilesPage, ProfilePage } from "pages/profiles";
+import LoginPage from "pages/login";
 
 const routes = [
   {
@@ -38,10 +39,10 @@ const routes = [
     component: () => "Rules",
   },
   {
-    name: "Ввійти",
+    name: "Увійти",
     exact: true,
     path: routesConstants.LOGIN,
-    component: () => "login",
+    component: LoginPage,
   },
   //   {
   //     exact: true,
