+++ src/styles/variables.module.scss
@@ -5,6 +5,9 @@
 $theme-c-black: #212121;
 $theme-c-white: #f3f3f3;
 
+$theme-f-playfair: 'Playfair Display';
+$theme-f-opensans: 'Open Sans';
+
 $shared-bp-tablet-portrait-up: 480px;
 $shared-iPad-portrait-up-min-width: 767px;
 $shared-bp-tablet-landscape-up: 979px;
